edit and delete logic are added

+++ src/main/java/me/akmaljon/rps/repository/GameRuleRepository.java
@@ -3,6 +3,6 @@ package me.akmaljon.rps.repository;
 import me.akmaljon.rps.entity.GameRule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameRuleRepository extends JpaRepository<GameRule, Integer> {
+public interface GameRuleRepository extends JpaRepository<GameRule, Long> {
     GameRule findFirstByFirstChoiceAndSecondChoice(String firstChoice, String secondChoice);
 }
